mcp-servers/example-server/requirements.txt
mcp
fastmcp
uvicorn
starlette
pyjwt
cryptography
python-dotenv
requests
